#1; moved directives to folders; created typeahead-result directive
