+++ projects/ngx-plug-n-play-lib/src/lib/ngx-plug-n-play-lib.module.ts
@@ -1,9 +1,8 @@
 import { NgModule } from '@angular/core';
-import { NgxPlugNPlayLibComponent } from './ngx-plug-n-play-lib.component';
 
 @NgModule({
-	declarations: [NgxPlugNPlayLibComponent],
+	declarations: [],
 	imports: [],
-	exports: [NgxPlugNPlayLibComponent],
+	exports: [],
 })
 export class NgxPlugNPlayLibModule {}
