+++ projects/ngx-plug-n-play-lib/src/lib/typeahead/typeahead-result.directive/typeahead-result.directive.spec.ts
@@ -1,0 +1,8 @@
+import { TypeaheadResultDirective } from './typeahead-result.directive';
+
+describe('TypeaheadResultDirective', () => {
+	it('should create an instance', () => {
+		const directive = new TypeaheadResultDirective();
+		expect(directive).toBeTruthy();
+	});
+});
