+++ projects/ngx-plug-n-play-lib/src/lib/typeahead/typeahead-result.directive/typeahead-result.directive.ts
@@ -1,0 +1,8 @@
+import { Directive } from '@angular/core';
+
+@Directive({
+	selector: '[libTypeaheadResult]',
+})
+export class TypeaheadResultDirective {
+	constructor() {}
+}
